+++ src/domain/providers/composer/serviceUtils.ts
@@ -54,11 +54,12 @@ export function addJsonClient(services: IServiceCollection) {
     serviceName,
     (container: IComposerService & IDomainServices) =>
       createJsonClient(
+        container.authorization,
+        container.authenticationSession,
         {
           caching: container.composerCachingOpts,
           http: container.composerHttpOpts
-        },
-        container.logger.child({ logGroup: serviceName })
+        }
       )
   );
 }