+++ src/infrastructure/infrastructureContainer.ts
@@ -1,4 +1,4 @@
-import { IServiceCollection } from '#domain/di';
+import type { IServiceCollection } from '#domain/di';
 import { addPackageFileWatcher } from './vscode/watcher/serviceFactory';
 
 export function addInfrastructureServices(services: IServiceCollection) {