+++ test/unit/extension/authorization/all.tests.ts
@@ -1,0 +1,3 @@
+export * from './urlAuthenticationSession.tests';
+export * from './authorization.tests';
+export * from './urlAuthenticationStore.tests';